refactor: inline conversation-events.js into talk.js
